Time: 134 ms (52.63%), Space: 44.6 MB (68.42%) - LeetHub

+++ 2882-ways-to-express-an-integer-as-sum-of-powers/2882-ways-to-express-an-integer-as-sum-of-powers.java
@@ -31,6 +31,6 @@ class Solution {
         // Option 2: skip this number
         int skip = dfs(remaining, idx + 1);
 
-        return dp[remaining][idx] = (int) ((include + (long) skip) % MOD);
+        return dp[remaining][idx] = (int) ((include +  skip) % MOD);
     }
 }
